+++ src/snowflake/snowpark/modin/plugin/extensions/series_extensions.py
@@ -253,7 +253,7 @@ def create_or_replace_view(
     name: Union[str, Iterable[str]],
     *,
     comment: Optional[str] = None,
-    index: bool = True,
+    index: bool = False,
     index_label: Optional[IndexLabel] = None,
 ) -> List[Row]:
     """
@@ -270,7 +270,7 @@ def create_or_replace_view(
             that specifies the database name, schema name, and view name.
         comment: Adds a comment for the created view. See
             `COMMENT <https://docs.snowflake.com/en/sql-reference/sql/comment>`_.
-        index: default True
+        index: default False
             If true, save DataFrame index columns in view columns.
         index_label:
             Column label for index column(s). If None is given (default) and index is True,
@@ -300,7 +300,7 @@ def create_or_replace_dynamic_table(
     data_retention_time: Optional[int] = None,
     max_data_extension_time: Optional[int] = None,
     iceberg_config: Optional[dict] = None,
-    index: bool = True,
+    index: bool = False,
     index_label: Optional[IndexLabel] = None,
 ) -> List[Row]:
     """
@@ -344,7 +344,7 @@ def create_or_replace_dynamic_table(
             - base_location: the base directory that snowflake can write iceberg metadata and files to.
             - catalog_sync: optionally sets the catalog integration configured for Polaris Catalog.
             - storage_serialization_policy: specifies the storage serialization policy for the table.
-        index: default True
+        index: default False
             If true, save DataFrame index columns as table columns.
         index_label:
             Column label for index column(s). If None is given (default) and index is True,
@@ -379,7 +379,7 @@ def to_view(
     name: Union[str, Iterable[str]],
     *,
     comment: Optional[str] = None,
-    index: bool = True,
+    index: bool = False,
     index_label: Optional[IndexLabel] = None,
 ) -> List[Row]:
     """
@@ -396,7 +396,7 @@ def to_view(
             that specifies the database name, schema name, and view name.
         comment: Adds a comment for the created view. See
             `COMMENT <https://docs.snowflake.com/en/sql-reference/sql/comment>`_.
-        index: default True
+        index: default False
             If true, save DataFrame index columns in view columns.
         index_label:
             Column label for index column(s). If None is given (default) and index is True,
@@ -426,7 +426,7 @@ def to_dynamic_table(
     data_retention_time: Optional[int] = None,
     max_data_extension_time: Optional[int] = None,
     iceberg_config: Optional[dict] = None,
-    index: bool = True,
+    index: bool = False,
     index_label: Optional[IndexLabel] = None,
 ) -> List[Row]:
     """
@@ -470,7 +470,7 @@ def to_dynamic_table(
             - base_location: the base directory that snowflake can write iceberg metadata and files to.
             - catalog_sync: optionally sets the catalog integration configured for Polaris Catalog.
             - storage_serialization_policy: specifies the storage serialization policy for the table.
-        index: default True
+        index: default False
             If true, save DataFrame index columns as table columns.
         index_label:
             Column label for index column(s). If None is given (default) and index is True,